Added Team, Role and TeamMembership models
The `TeamMembership` model keep the relationship between teams, users and
roles.
A user belongs to a team with a given role (e.g. "maintainer" or "member").

A user can be in a team only once and only with a single role:
  * add a user to a team more than once doesn't make any sense
  * add a user to a team with more than one role is confusing and doesn't
    make any sense (e.g. has the user the "member" role or the "maintainer"
    role?)

I'll add the following methods to Team and User to make the API a bit more
user friendly:
  * `User#teams` to return the teams a user is in
  * `Team#users` to return the users in a team
  * `Team#users_with_role` to return the users in a team with a certain
    role