Create/Delete AWS IAM roles for apps
- When an app is created we create its role in AWS.
- The app role name would follow this pattern: `$ENV_app_$APP_NAME`
- When an app is deleted we delete the corresponding IAM role
- IAM roles will be assumed by the app's k8s pods this means the role
  assume policy needs to contain the k8s worker role ARN as principal
  (See [`sts:AssumeRole` required by kube2iam](https://github.com/jtblin/kube2iam#iam-roles))
- This ARN is passed in the settings via the `K8S_WORKER_ROLE_ARN`
  environment variable
- I'll need to update the cpanel helm chart and the cpanel helm chart config
  to pass this setting in order for this to work

* Related code

This is somewhat similar to what we already to in the users lamba functions:

https://github.com/ministryofjustice/analytics-platform-ops/blob/19046f0c10da4cde551bdc6f0847e332b0d92a7d/infra/terraform/modules/data_access/users/users.py#L18

* Notes regarding the tests/design decisions

The code is mainly split in 3 layers:

`views` <=> `services` <=> `aws`

Each layer's tests only test that the right calls to the adiacent layer
are perfomed correctly.

The tests only test the "public" functions treating them as "black boxes"
and testing that given a certain input certain side effects happen.

As a concious design decision the `aws._detach_role_policies()` is "private".
This is (at the moment!) only used internally by `aws.delete_role()` when
(and if!) there will be the need to call it from the outside we'll make it
"public", until that moment this is an internal implementation detail
(which is of course tested) but not leaked into the client code.

**IMPORTANT**: We should try to find a way of mock every boto3 call while
running the tests as a precautionary measure: as it stands it's possible
to accidentaly write tests that could potentially interact with the real AWS
(which is bad). At the moment I renamed my AWS credentials locally to be sure
in this event boto3 would simply raise a "missing credentials" exception.
This is not ideal and we should discuss a possible solution for this.

* Ticket

https://trello.com/c/AULNJYns/359-add-iam-role-for-control-panel-api-app-to-terraform