Added 'UserS3Bucket' model
This is the many-to-many relationship model between users and S3 buckets
which will determine which buckets users will have access to.

As this conceptual is **very** similar to AppS3Bucket I factored out the
common parts in `models.AccessToS3Bucket` which is an *abstract* model.

The concrete subclasses (`AppS3bucket` and `UserS3bucket`) will only add the
specific fields and functionality and inherit the common fields and the rest.

The `UserS3bucket` model also has a `is_admin` boolen field (default `False`)
which will determine if the user has "admin" privileges, therefore can
grant access to other users/apps.

Ticket: https://trello.com/c/xLiTXJV8/361-add-relation-between-users-and-s3-buckets-to-control-panel-api