Generate App slug (used to generate IAM role name) from 'repo_url'
As discussed with rest of the team we want apps' IAM roles to be
generated from the repository URL, rather than from the app name
which is entered by the user.

This will allow Jenkins to be able to generate the IAM role name
for the app from the repository URL (which is available to it)

The repository "name" is the part of the repository URL between
the last slash (`/`) and `.git` (if present).