Updated `K8sProxy` to load cluster configuration
Will use `kubernetes` module to try read the configuration from the cluster
the app runs on first. If the app is not running in a k8s cluster (e.g.
running the app locally) then will try to read kubectl configuration from
file - this means will proxy requests to whichever cluster is currently
in use locally.