Proof of concept of `Tool` class to deploy...tools
We currently only have RStudio as a tool but we want to add more.

A tool as a canonical name (e.g. 'rstudio' or `jupyter`) and can be
deployed for a given user. This will upgrade (or install if 1st deploy)
the given tool in the user's k8s namespace.

The right helm chart is determined using the tool name.

 - The Auth0 client id/secret are read from `${TOOL}_AUTH_CLIENT_ID` and
   `${TOOL}_AUTH_CLIENT_SECRET` respectively
 - The client domain is read from `settings.OIDC_DOMAIN` (as this should,
   at least for now, be the same for all clients)

Usage example:

```python
>>> rstudio = Tool('rstudio')
>>> rstudio.deploy_for(alice)
```

NOTE: Regarding the Auth0 client config, we may want something more clever
      in the future but for the short-medium term read this from the
      environment variables is fine. We can iterate on this.