K8s: Fix permissions to disallow requests to namespace with common prefix
Because of the use `startWith()` to check the k8s request path is for the
user's namespace and the lack of trailing slash `/` a user could potentially
make request to operate on someone else namespace.

This happens if the user has a username which is prefix of another user'
username, e.g.

- User `andy` has a namespace called `user-andy`
- User `andyother` has a namespace called `user-andyother`
- `andy` should only be able to operate in `user-andy`
- `andy` can operate in `user-andyother` because this namespace starts
  with `user-andy`.

Ticket: https://trello.com/c/4zkAt4r3/556-1-cp-api-improve-k8s-namespace-permissions