Moved tests for UserPermissions under tests/permissions
Having a test file per class is cleaner and also has the benefit of
reducing the chances of conflicts when people are working on relatively
close parts of the code (e.g. permissions for `S3Bucket` and `UserS3Bucket`)

I didn't move everything as I don't want to move code potentially currently
being modified by someone  else but long term this is the way I want tests
to be structured (especially when test files start to get massive with
500+ lines of code).

Potential future candidates for this refactoring are tests for filters and
views (and maybe others).