Migration to add S3Bucket records for buckets created by lambdas
These were created when the corresponding GitHub team was created.
This migration only creates the records in the DB and set the
"data warehouse" flat to true.

The migration starts from IAM policies as this way we can ignore
policies which are not attached to any role (e.g. not used).

Migration only consider policies starting with `{ENV}-`, e.g. `dev-`
and ending with `-readwrite` (as lambdas only use these policies).

It also ignores potential app buckets (e.g. `dev-app-`) as these are
already handled by CP-API and they're not data warehouse buckets anyway.

Part of ticket: https://trello.com/c/Me2RD6ek/696-5-script-that-copies-data-access-permissions-from-existing-lambda-created-policies-to-to-s3-access-inline-policy-and-adds-these