Normal users can list users
Updated the User permissions/filtering to allow normal users to list all
users. This is a requirement coming from the Clive/UI.

The reason for this is to allow normal users to grant access to other users
to S3 bucket they are admin of.

The changes are in two areas:
- Users are not filtered anymore, e.g. all users are returned by `UserFilter`
- `UserPermissions` now allows `list` actions for everyone. Note that normal
  users will still only be able to retrieve/update their own user record only.

**NOTE**: I inverted the logic in `UserPermissions` to be a whitelist of
actions rather. It's easier to see what a normal user can do and also
generaly whitelists are safer.

Ticket: https://trello.com/c/ivwsh6FI/715-3-normal-users-can-list-users