When deleting an S3 bucket, update the permissions on associations
Permissions in DB are kept in two models `UserS3Bucket` and `AppS3Bucket`.
These records contains a reference to the S3 bucket.
When these records are deleted the user/app IAM role needs to be updated
to reflect the fact that these users/apps don't have access to the
corresponding S3 bucket.

This happened when these records were deleted directly but not when deleted
inderectly when the deletion of an S3 bucket cascaded and deleted all the
associated records.

As we don't actually delete the S3 bucket in AWS a user/app would still
have access to that "deleted" S3 bucket.

Solution is to listen to the `post_delete` signal which is [still triggered](https://stackoverflow.com/questions/8944207/overridden-delete-method-of-child-model-does-not-get-called-when-deleting-parent)
when a record is deleted by the cascading.

NOTE: As update the IAM roles when access to S3 bucket is revoked is now
performed by the signal, I deleted the custom `perform_destroy()` in the
views which were added to do that.

Ticket: https://trello.com/c/f0gPNADq/720-4-delete-data-source-not-updating-permissions