When a user logs the 1st time, create their IAM role
When a user logs for the first time, the `authentication` module
detects that the user is not in the DB yet and creates it.
It also install the `init_user`/`config_user` helm charts for the user.

It now also creates the user IAM role as well (as the `POST /users`
endpoint already does).

This was not a problem before before users' IAM roles were created
by an AWS Lambda function when the user was added to the GitHub org.
As we'll move away from that model, it's important that the CP
creates these IAM roles.

NOTE: I manually triggered the `User.aws_create_role()` in `dev` on a
user which already has the IAM role and it worked fine. e.g. this
is idempotent.

Ticket: https://trello.com/c/Op8lIxJi/740-2-iam-role-not-created-when-user-logs-first-time