Fixed jsConfirm for `<button>` elements
The class selector was only handling `<input type="submit">` but not
`<button>`. For example the "delete data source" button was a `<button>`
with `.js-confirm` class but it didn't ask for confirmation.

I've tweaked the module so that now doesn't care of which element the
`.js-confirm` class is attached as long as it's a children of a `<form>`
it would ask for confirmation before submitting the form.

**NOTE**: Normal links (`<a>`) with `js-confirm` class are still handled
as before.