Added ability to "Upgrade" a tool
This is functionally the same as deploying it (the command is the same).
We only show the "Upgrade" button when a user deployed version of a tool
has a different version of the version in the DB (which is the version
we consider stable and supported).

Ticket: https://trello.com/c/rz6o5bug/101-users-can-upgrade-their-tools