Add css / babel watch scripts to package.json
This commit allows you to run `npm run watch` which will keep compiling assets
when files are changed so you get quicker feedback when developing front-end
features.

I've added `corejs@2` to the requirements to get rid of a babel warning, and
delegated some of the makefile tasks to just call into `npm run` for css / babel.