k8s request to get App URL is made using CP service account
This is a necessity as users only have permission to operate in their
own namespace and they shouln't have to get resources in `apps-prod`
(where the apps live).

As a result I've added a `use_cpanel_creds` parameter to the
`KubernetesClient` class.

(I was planning to do some further simplication but it turned out to be
more involved than expected and left it for a standalone refactoring PR)

`controlpanel/api/cluster.App.url` can be a property as we don't need
to pass in the `id_token` anymore.