Raise exception when get_id_token() can't find a request
This could happen if for whatever reason (bug) we call
`models.User.get_id_token()` in a background worker where there are
no requests.

This will raise an exception and highlight the problem straight away
(hopefully while running the tests but potentially in production).

Most importantly it wouldn't continue the execution potentially leading
to strange behaviour or even security problems. For example instanciating
`KubernetesClient` with blank `id_token` would default to use CP
credentials which is not only unexpected but also potentially dangerous.