Fix for KubernetesClient error when no id_token provided
That check was in place for some refactoring I'm working on (which
should remove the code that reads the ID token from request when not
passed).

The problem is that currently some of the code don't pass the `id_token`
and let the class read from current request.

I'm commenting that check for now.