Workaround for race condition on kubernetes credentials
This was causing some strange `401` but not all the times.

Description of the problem
==========================
To get a working kubernetes configuration from the `kubernetes` module
you usually call `load_kube_config()` or `load_incluster_config()`.
The config is "loaded" and can be retrieved by instantiating
`kubernetes.config.Configuration()` which set that as default configuration
and then return a shallow copy (not a deep copy) of it.

Unfortunately some of the attributes are dictionaries which means that
they'll be shared across all instances of a `Configuration`.

When we instanciate a new kubernetes client and set `api_key` and
`api_key_prefix` these will be overwritten in the single dictionary shared
across all the `Configuration` instances - which is bad.

Solution
========
When a new config is loaded and retrieved, we make a copy of the `api_key` and
`api_key_prefix` keys so that they're not shared.

Probably related to https://github.com/kubernetes-client/python/issues/932

Ticket: https://trello.com/c/D4bpabeD