Removed legacy `ENABLE_K8S_RBAC` flags
This was useful when we enabled RBAC in the k8s cluster and we
transitioned to it but it's not necessary anymore as this is always
enabled now.

It also works fine when running the app locally and tests pass so no
reason to keep this.

The main reason to remove this is that it will enable us to simplify
the `KubernetesClient` class and make it more testable (as we'd not
need to care about testing cases when this flag is enabled or disabled)

Part of ticket: https://trello.com/c/OXWwaZua