Always pass `id_token` explicitly
Some of the functions in the `cluster` module and some of the methods in
the `ToolDeployment` "model" were quite flexible regarding the `id_token`
argument.

This now changed and if the client code wants to perform an operation
against the k8s cluster it needs to provide an `id_token`.

I've also removed the `_id_token` status in `models.ToolDeployment` as
it was confusing, sometimes it was set but used later.

This should allow us to remove the logic in the `KubernetesClient` class
that automagically reads the `id_token` from the request when not passed
(this can potentially lead to bugs if an instance is created in a context
without a request)