IAM role creation works in cross-account setup
A quirk of AWS cross-account setup is that the IAM role trust relationship
(AKA `AssumeRolePolicyDocument`)'s `Principal` is not the full ARN of the
assuming IAM role but only the ID of the account where the assumimg IAM
role is.

Once we have both data and compute AWS account IDs we can then tell
if everything is in the same AWS account or not (e.g. different IDs means
we have a cross-account setup).

In practice this means that CP would work as before in `alpha` (as long
as `AWS_COMPUTE_ACCOUNT_ID` and `AWS_COMPUTE_ACCOUNT_ID` as set correctly
to the same account ID where everything sits).

**IMPORTANT**: Remove now-ambiguous `AWS_ACCOUNT_ID` and added
`AWS_COMPUTE_ACCOUNT_ID` and `AWS_COMPUTE_ACCOUNT_ID` configuration.
We need to distinguish between the AWS account where the data/IAM roles
are and the AWS account k8s nodes (tools and apps) are.

Part of ticket: https://trello.com/c/1DrJ8TN1