Fixed revoke access when IAM role doesn't exist
When trying to revoke access to a bucket the code tries to load the
corresponsing IAM role's `s3-access` inline policy but if the IAM role
doesn't exist this would of course fail.

It's not clear what's causing this but this shouldn't prevent the operaion
to succeed as if there is no IAM role effectively the provided IAM role
(which doesn't exist) already doesn't have access to the bucket and hence
there is nothing to revoke.

Also, as when deleting a user/app role all the associated `AppS3Bucket`
and `UserS3Bucket` records are deleted this exception mean you can't
delete a user/app when it has access to something but no IAM role for
whatever reason.

Ticket: https://trello.com/c/vwgmnhBX