_tag_bucket() helped keeps existing tags
`Tagging().put()` overwrite existing tag set. This is not ideal as we
may leverage existing tags in the archiving process (e.g. we may
simplify archiving logic by using also other tags like the `env` tag)

This may also be useful to avoid a situation in which a bucket is created
in CP just after an old bucket was tagged `to-archive` but not archived
yet. If CP overwrites tags and `to-archive` tag is lost the bucket and
its data may not be archived as intended.

Above case shouldn't really happen right now because of how code is
currently written but future refactoring may risk to change this.

Basically the `_tag_bucket()` intended behaviour is to add tags rather
than overwrite them.