When a bucket is deleted, tag it with `to-archive=true`
This will instruct the [`buckets-archiver`] to archive its data and
then delete the S3 Bucket in AWS.

[`buckets-archiver`]: https://github.com/ministryofjustice/analytics-platform-buckets-archiver

One technical subtletly is that I'm overriding the `models.S3Bucket#delete()`
instance method to also make an AWS call to tag the bucket (via the
`cluster.S3Bucket#mark_for_archival()` method).

This operation is wrapped in a DB transaction to be sure that if for
whatever reason the API call to AWS fails (e.g. network connectivity,
permissions, anything) the S3 bucket record is *not* deleted from the DB
as that would leave the S3 bucket untagged for archival.