auth0: refactoring to add get_group_id()
All existing code using `get_group()` were only getting the Authz
group to get hold of the group's ID. This refactoring makes this
explicit.

The `get_group_id()` method is still reusing the `get_group()` method.

**NOTE**: Currently the `get_group()` method loops through *all*
the Auth0 groups just to get the group we're interested in.
This is because Auth0 API doesn't have an API to get a group by
its name.
This is very inefficient (especially as the number of Authz groups
grows) and ideally we'll want to improve this to cache the Auth0
Authorization (authz) API groups' IDs. This shouldn't be a problem
as they're not supposed to change. Potentially we could either load
them into a cache on app start or store them in a field of the
App model/table and maybe populate these fields with a migration
and/or lazy-load them in the `auth0.get_group()` method (e.g.
look for the group_id in the DB first, and if not make API call and
and update app record. Although this is a bit messy)