On app delete, also delete Auth0 roles and permissions
This needed a bit of rework because of the way the Authorization
Extension API works: Delete a role or permission when there are
associated groups or roles throws a `400 BAD REQUEST`.
This means we need to get the roles IDs and permission IDs and
delete things starting from the group ("top of the tree"), then down
delete the group and finally the role.

This was also tested manually and it worked (on top of the usual
automated tests)