Tweaked `App.delete()` test
I've simplified it and it now just checks if it's calling the
`cluster.App().delete()` method instead of checking all the various
operations performed on delete.

This is sufficient as the `cluster.App.delete()` method is tested already
so no need to double-check the internal implementation details in the
test for the model. The App model only needs to know that the method
to delete the app from the cluster is called and trust that method
to perform what is needed.