Better handling of missing app Auth0 group
We can't add customers to an app until the Auth0 Authorization
API group is created.

Raising an exception as soon as we know the group is not found.
No point in continuing trying to add customers to a group
when it's not there.

This group is created by Concourse when the app is released
the first time.