use battle-tested version of kubernetes (`10.0.1`)
One of the tests highlighet that `kubernetes==11.x` has an API
change which breaks some of our code.

I tried to upgrade only the minor version to `kubernetes==10.1.0` but
for some reason this newer version downgrade/requires `pyyaml~=3.12`:

```
kubernetes 10.1.0 has requirement pyyaml~=3.12, but you'll have pyyaml 5.1.2 which is incompatible.
```

So in the end I've sticked to `kubernetes==10.0.1` for the time being.

Part of ticket: https://trello.com/c/RtN2pCxQ