Uninstall old RStudio release before helm upgrade
Due to the change in the release naming scheme of RStudio, old users
can't upgrade their RStudio from the Control Panel because of an
underlying helm error.

This can be manually resolved by running `helm delete --purge` on
old release but it's annoying from a support and UX standpoint.

After this change the CP will check if there is a release with
the old naming scheme (e.g. `USERNAME-rstudio` instead of the correct
`rstudio-USERNAME` name) and delete it before installing the new release.

Part of ticket: https://trello.com/c/DBJhJKPT