Made some of `Helm`'s methods class methods
AFAICT none of the methods in the `Helm` class actually **need** to
be instance methods, none of them seem to use `self`.

I kept this refactoring light and only changed `_execute()` ("private")
and `update_reposoritories()` (only used within the `Helm` class).

I want to play with the idea of a separate `HelmRepository` class
responsible for the update of the helm repository and possibly
keeping tabs on charts versions as part of the broader "tools catalogue"
work.

Related to: https://trello.com/c/DBJhJKPT