HelmRepository: Ability to get chart info/versions
This `HelmRepository.get_chart_info()` class method allow the client
code to retrieve information about all the available chart releases
and the corresponding tool version.

This is to answer questions like "What version of RStudio/R was shipped
with the rstudio-2.0.0 chart?".

This information is retrieved by `helm update` and this method reads
and parse the YAML file with the repository index.

This read/load operation only happens if there isn't a relatively
recent version of this index (30m) and prevent several `helm update`
run and corresponding `YAML.load()` to happen when not necessary.

Part of ticket: https://trello.com/c/74QeufKf