Tools-related refactorings
Added few methods and hopefully simplified things a little
by adding few methods for operations which were previously
performed inline and could be factored out:

- Added `cluster.ToolDeployment.get_installed_chart_version()`,
  this is to get the version of a tool chart currently installed
  for the user (based on `Deployment`'s `chart` label currrently)
- in `ToolDeploymentManager.filter()` I've removed the code which
  was simply used to get the current chart version which was then
  passed to the `models.ToolDeployment` constructor.
  The aim of this was really to determine if a chart needed upgrade,
  this is now achieved by using the `models.ToolDeployment.outdated()`
  method
- again, the `models.ToolDeployment` constructor doesn't receive the
  `deployed_chart_version` argument anymore as this was simply used
  to determine if the installed chart was outdated (now we have the
  `outdated()` method to determine this)
- renamed `models.ToolDeployment.deployed_tool_version()` to
  `get_installed_app_version()`, the name may be better but it
  includes `app_version` to highlight the fact this is meant to return
  the `appVersion` field from the helm repository.
  This method now also uses the `get_installed_chart_version()` (
  `cluster.ToolDeployment`) and needs an `id_token`.
  It's a nice inconvenience as this method will likely be re-used in
  the backgroud tasks/`consumer` module which need to explicitly
  pass the `id_token` to interact with kubernetes
- again, the `models.ToolDeployment.outdated` method is now using
  the `get_installed_chart_version()` (`cluster.ToolDeployment`), this
  requires an `id_token` (as it interacts with the kubernetes cluster)
  but this way the `ToolDeployment`'s constructor doesn't need to
  receive the current version of the installed tool chart which was
  hacky.

This is all kinda preparation work related to the Tools Catalogue
and specifically this ticket: https://trello.com/c/74QeufKf

**NOTE**: Current implementation of these methods to determine current
version relies on the `chart` label of the `Deployment`, this is been OK
and it seems to work so far so I'm not going to change it risking
introducing bugs.
Potentially we could retrieve the same information using `helm list` in
conjunction with the `--output=json` parameter but that has its set of
challanges.
Furthermore above would require changing the `Helm.list_releases()` method
as it currently uses the `-q/--short` flag to return only the release names,
and it use `strip` to build a list. Switch to JSON will require few
tweaks and at this stage may be overkill.