Added some tests for the frontend.consumers module
There is some repetition in these tests but considering there were
no tests at all before this is somewhat an improvement.

We could factor some of the logic out but this module will likely
change soon so I thought it was worth adding some tests even if not
perfect until we move things around further.