Tools Deploy/Upgrade honour version in Tool record
Before this change the Control Panel was a bit "unpredictable" regarding
the version of a tool the user would get:
helm upgrade/install would not specify the version and the latest
version would be installed.

This was not too bad but as we improve the Tools management part of
the Control Panel this needs to be main more specific.

This is preparation work so that when a user tries to upgrade a tool
and we say "This will upgrade it to version x, y, z, OK?" and the user
confirms they don't then get a different version.

Related to ticket: https://trello.com/c/jHI3iCHq