Show app version on tool Deploy/Upgrade
When a user click on the Deploy or Upgrade button the Control Panel
now shows them the version of the tool they're about to install or
upgrade to.

This version is actually the `appVersion` information coming from
the helm chart, this should contain useful information, for example
for RStudio would be something like

> RStudio: 1.2.1335+conda, R: 3.5.1, Python: 3.7.1, patch: 10

This will give the user enough details to take an informed decision
on whether to install or more importantly upgrade.

NOTE: As part of this change I've moved the logic to get a tool/chart
`appVersion` into the `HelmRepository` class as it's now used in more
than one place and it makes sense for this to be in the class used to
query the helm repository.

Ticket: https://trello.com/c/jHI3iCHq