Formatted some more Python code with black
We currently don't format all the python code automatically but we
think it may be a good idea to do it more and more.

Eventually we'll get to a point where all the code has the same coding
standards (although some of the formatting decisions may make the code
slightly less readable, e.g. see `test_tool.py` file in this commit)

NOTE: Keeping these formatting changes in separate commits for the
moment to make it explicit in Git history. Would help when investigating
bugs with with Git bisect/blame and reduce the time spent getting to the
real commit which may have caused it and the reasons why these changes
were made.

Keeping these formatting changes with other changes may confusing for
the person investigating.