Updated ToolList view to support multiple versions of a ToolList
Before this assumption was that there would be at most 1 instance
of Tool for each of the tools. This to signify the fact we only
support one specific version of a tool's chart.

As part of ticket: https://trello.com/c/DBJhJKPT related to the
Tools Catalogue we aim to allow users to chose the version of the chart
to use among few supported ones. Although for some tools we may still
only have one supported version.

This means that potentially a user may be able to chose **for example**
between something like "RStudio 1.2 with R 3.5" and "RStudio 1.3 with
R 3.9" (these are not real versions).

This means that the tools DB table may contain two records for the
`rstudio` helm chart and the tools template needs to be aware of this.

This commit arrange these information into the `tools_info` variable
in the context for the `ToolList` view. I've experimented with the UI
and this data should allow it to give the users the options.

But the UI is still work in progress and it's not part of this commit.

(there are some UX challenges there as we move from a button to "Deploy'
and a button to "Upgrade" into something more complex with a `select`
showing the appVersion (which is a long string), especially as current
logic to update page on status events will have to change as well)