Removed code related to Tool "Upgrade"
This will not be used anymore as there is nothing really special
about the concept of "Upgrade", in practice this is installing an
helm chart (now specifying the version) and it's effectively the
same as the "Deploy" operation.

Furthermore, in order to allow users to select a tool version from
a shortlist of supported version the UI will have to change to some
sort of select and therefore the user would select a version and then
click "Deploy" to install it. This means there is no distinction
between first installation and subsequent upgrades.