Removed `models.Tool.outdated()`
This method was mainly there to determine whether to show or not the
"Upgrade" button which is now gone.

Furthermore, this method doesn't make sense anymore as there may be
two records for the same chart with different versions, e.g.

`v1 = Tool(chart_name="rstudio", version="v1")` and
`v2 = Tool(chart_name="rstudio", version="v2")`

so what would it mean to call `v1.outdated()`?

The new model where the user can simply see the list of versions for a
tool and select/"Install" the one they want is simpler and sidestep
the need for this method.