Removed unused 🤞🏼ToolDeployment view and related
This view was there but as far as I can tell it was not used.

The `ToolDeployment` view had its own URL, permissions etc...and
it supported create and list to deploy a tool for the current user and
to list the ToolDeployment for a user.

But AFAICT this view is not used, the Tools page is using the endpoints
provided by the `controlpanel.frontend.views.tool` module.
These view submit a background job processed by the `consumers` module.

As this doesn't seem to be used I was able to remove the view, the
serialiser, the permissions classes etc...

An even nicer consequence of removing this view is that I was able to
delete the weird/strange `ToolDeployment.filter()` method which always
felt wrong (it was aggregating informations from the Tools DB table and
the Kubernetes deployments and wrapping/returning them as a list of `ToolDeployment` instances)

This method was previously also used in the `ToolList#get_context_data()`
method (`frontend.views.tool`) but it's not anymore and hence the removal.

(the `ToolList` view still aggregate information from the Tools DB table
and the user k8s deployments for tools but it's doing it in a more explicit
and hopefully less confusing way)

NOTE: The code removed seem to be only used in tests which were failing
that's why these tests were deleted as they were testing unused functionality