Moved (temporary) Tools list CSS into stylesheet
This was inline as I was trying to make the page look pretty enough
and in anticipation for this to be experimented/improved on further.

Having said that I now moved this into a stylesheet as it should.
If the page markup will change and it will no more necessary we can always
delete.