Fix Dockerfile with python version bump
This fixes an error on Dockerfile build:
```
ERROR: unsatisfiable constraints:
  python3-dev-3.7.7-r1:
    breaks: world[python3-dev=3.7.7-r0]
```
which is no doubt caused by the base Alpine packages now coming with 3.7.7-r1 now.

(We might also consider specifying <3.8, to take the latest 3.7.x release. But on balance we
might take a bit of brittleness since we value pinning/reproducibility)